pyproject.toml
[project]
name = "coloco"
version = "0.3.0"
description = "A kit for creating FastAPI + Svelte applications"
authors = [{ name = "Channel Cat", email = "channelcat@gmail.com" }]
dependencies = [
    "aerich[toml]==0.8.1",
    "fastapi==0.115.0",
    "PyJWT==2.10.1",
    "tortoise-orm[asyncpg]==0.24.0",
    "type-less==0.1.4",
    "typer==0.15.2",
    "uvicorn==0.31.0",
]
requires-python = ">=3.11"
readme = "README.md"
license = { text = "MIT" }

[project.scripts]
coloco = "coloco.__main__:app"

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true
